moved the feed algo to a new file. New file for LocationViewModel.
